20. Big Changes to EscrowFactory and EscrowExchange to Make Contracts Compile Properly
All the changes made today were with the idea that we typically use with MySQL and how we modify current records in applications with databases.

We used a contract_id to identify a specific record to retrieve or modify (i.e. CRUD – Create, Read, Update, Destroy) except we really only do CRU when it comes to blockchain development.

But the question stands "Will the old way of doing things like we did in Web 2.0 work in Web3.0?"

You'll find out in coming episodes.