23. Learning Truffle Tests
Quick rundown on testing with Truffle. Testing is incredibly important and most of the time, it's the only way to test that your code actually works.
We will be going back to tests a lot in future episodes, so don't worry if this blows over your head. Remember, you're learning by doing.

That said, below are some resources for honor roll students who want to learn on their own.

Resources:
https://www.trufflesuite.com/docs/truffle/testing/writing-tests-in-javascript
https://gist.github.com/yoavniran/1e3b0162e1545055429e#chai
https://www.chaijs.com/api/assert/