24. Adding Notes to Our Escrow Contracts
We will show you how to add notes to your contract when you create it. Doing so will require us to modify our frontend App.js and Main.js to something that
can handle that.

We'll also look into using react-bootstrap, which I won't go into detail here, so for homework make sure to go see the resources below and familiarize yourself
with react-bootstrap because that is used pretty much everywhere.

Resources:
https://react-bootstrap.github.io/