25. Demo of New UI and Adding DepositCount and AmountCount
Not many crazy changes. Just adding depositCount and amountCount, which I'm not even sure we would want at the end of this project. But again, it's
a way of trial and erroring and testing the whole process, because that's how you learn.

If you're UI isn't compiling, check the following:
1. Delete your build folder and re-run "truffle migrate"
2. Make sure your App.js is updated (i.e. we got rid of "Escrow" as mentioned in the video)
3. Make sure that your EscrowExchange.json and EscrowFactory.json files have "networks" (refer to video)
4. If that all doesn't work, scream for help?