29. REDOING EscrowFactory and Contract Management!?!
We are NUKING EscrowFactory and REDOING everything. WHY? I'll explain in the video. Just stick with me. It's part of coding,
sometimes you have to nuke your hard work in order to start fresh to build on something simple rather than complicated because
building on complicated things WILL RESULT in a mess later!

I'll make this as painless as possible. Let's just get through it!