32. Iterating through Contracts Efficiently
Alright, the example here is we have a mapping of a bunch of contracts like so:

1 2 3 4 5 6 7 8 9 10
A B C E C A E Z U T

Now, on the front end, we don't want to iterate through 10 places for our entire array when looking for all of A's contracts. We just want to look for the first 6
because 6 is the last contract for A. How do we do that? That's what we'll explore in today's episode.