35. Big Refactoring of Tests
So, we did some moving around of our tests. Make sure to grab the latest changes from the commit that accompanies this video and copy that to your local project.

HOMEWORK:
Spend an evening and try to fix broken tests (not all will pass). When you're tired of it. Go to the next episode.