43. FINAL PROJECT DEMO
Fix the package.json problem and rebuild node_modules if you need to.
THE BEST FIX is to go get the package.json from the commit that accompanies this video and copy and paste it in your local project.
DON'T FORGET to delete package-lock.json and node_modules before running `npm install`

DON'T FORGET to delete the build folder and re-run `truffle migrate` to get the updates on your local blockchain on Ganache